server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
websockets>=14.0
aiosqlite>=0.20
openai>=1.60
anthropic>=0.42
google-genai>=1.0
apscheduler>=3.10,<4
python-dotenv>=1.0
pyyaml>=6.0
httpx>=0.27
